fix(spec_decode): upcast token/req indices to int64 in sampling kernels
Upstream vllm #46560 fixed a silent out-of-bounds access when the cumulative token count (num_logits) exceeds INT32_MAX: triton-ascend computes pointer offsets in int32, so token_idx * logits_stride overflows and wraps around at large token counts. Apply the same int64 upcasts on NPU:
- gumbel.py: _temperature_kernel and _gumbel_sample_kernel cast tl.program_id(0) and the expanded_idx_mapping load to int64 before pointer arithmetic.
- rejection_sampler_utils.py: _probabilistic_rejection_kernel casts the idx_mapping / cu_num_logits loads to int64.
- resample.py: _resample_kernel and _categorical_finalize_kernel (the resample path was moved there and rewritten in #15573) cast the cu_num_logits / expanded_idx_mapping loads to int64; req_state_idx must be int64 because it multiplies the draft logits stride (max_num_reqs * num_speculative_steps * V can exceed INT32_MAX).

The original _npu_gumbel_block_argmax / _resample_kernel hunks target kernels that no longer exist after the #15573 rewrite; the equivalent offsets in the new kernels get the same casts.

+++ vllm_ascend/ops/triton/v2/spec_decode/resample.py
@@ -60,10 +60,13 @@ def _resample_kernel(
         req_idx = task_idx // num_blocks
         block_idx = task_idx - req_idx * num_blocks
         resample_idx = tl.load(rejected_step_ptr + req_idx)
-        start_idx = tl.load(cu_num_logits_ptr + req_idx)
+        # Cast to int64 before pointer arithmetic: triton-ascend computes
+        # offsets in int32, which overflows when num_logits exceeds INT32_MAX
+        # (upstream vllm #46560).
+        start_idx = tl.load(cu_num_logits_ptr + req_idx).to(tl.int64)
         end_idx = tl.load(cu_num_logits_ptr + req_idx + 1)
         resample_token_idx = start_idx + resample_idx
-        req_state_idx = tl.load(expanded_idx_mapping_ptr + resample_token_idx)
+        req_state_idx = tl.load(expanded_idx_mapping_ptr + resample_token_idx).to(tl.int64)
         temperature = tl.load(temp_ptr + req_state_idx).to(tl.float32)
         is_bonus = resample_token_idx == end_idx - 1
         needs_resample = (temperature != 0.0) | is_bonus
@@ -145,10 +148,13 @@ def _categorical_finalize_kernel(
     """Select the final token using one global categorical threshold per request."""
     req_idx = tl.program_id(0)
     resample_idx = tl.load(rejected_step_ptr + req_idx)
-    start_idx = tl.load(cu_num_logits_ptr + req_idx)
+    # Cast to int64 before pointer arithmetic: triton-ascend computes
+    # offsets in int32, which overflows when num_logits exceeds INT32_MAX
+    # (upstream vllm #46560).
+    start_idx = tl.load(cu_num_logits_ptr + req_idx).to(tl.int64)
     end_idx = tl.load(cu_num_logits_ptr + req_idx + 1)
     resample_token_idx = start_idx + resample_idx
-    req_state_idx = tl.load(expanded_idx_mapping_ptr + resample_token_idx)
+    req_state_idx = tl.load(expanded_idx_mapping_ptr + resample_token_idx).to(tl.int64)
 
     temperature = tl.load(temp_ptr + req_state_idx).to(tl.float32)
     is_greedy = temperature == 0.0
